[TLC-249] Update integration tests for IdentifierRestRepository
